cleaning up code for readability
